- adaptation for the kernel strdel(); - buffer overrun removed (rare crashes); - old useless code removed; - removal of a blank line after comments was a bad idea
git-svn-id: http://svn.miranda-ng.org/main/trunk@15136 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/msgdlgutils.h
@@ -64,7 +64,7 @@ int    TSAPI MsgWindowDrawHandler(WPARAM wParam, LPARAM lParam, TWindowData *dat
 void   TSAPI LoadOverrideTheme(TContainerData *pContainer);
 void   TSAPI LoadThemeDefaults(TContainerData *pContainer);
 void   TSAPI ConfigureSmileyButton(TWindowData *dat);
-int    TSAPI CutContactName(const TCHAR *szold, TCHAR *sznew, unsigned int size);
+int    TSAPI CutContactName(const TCHAR *szold, TCHAR *sznew, size_t size);
 void   TSAPI SendNudge(const TWindowData *dat);
 void   TSAPI EnableSendButton(const TWindowData *dat, int iMode);
 LRESULT TSAPI GetSendButtonState(HWND hwnd);
